Fix ValueError on AuthorAlias search
